backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-multipart==0.0.20
pydantic-settings==2.7.0
tensorflow>=2.20,<2.22
python-docx==1.1.2
pdfplumber==0.11.4
pytest==8.3.4
httpx==0.28.1
reportlab==4.2.5
